+++ docs/source/conf.py
@@ -20,6 +20,7 @@ extensions = [
     "sphinx_design",
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
+    "sphinx.ext.autosectionlabel",
     "sphinx.ext.intersphinx",
     "sphinx.ext.napoleon",  # parameters look better than with numpydoc only
     "numpydoc",
@@ -29,6 +30,8 @@ extensions = [
 autosummary_generate = True
 # dont show __init__ docstring
 autoclass_content = "class"
+# for uniqur labels/anchors
+autosectionlabel_prefix_document = True
 # sort class members
 autodoc_member_order = "groupwise"
 # autodoc_member_order = 'bysource'