+++ src/finam/sdk/adapter.py
@@ -8,13 +8,8 @@ from typing import final
 
 from ..data import tools
 from ..data.tools import Info
-from ..interfaces import (
-    FinamLogError,
-    FinamMetaDataError,
-    FinamTimeError,
-    IAdapter,
-    IOutput,
-)
+from ..errors import FinamLogError, FinamMetaDataError, FinamTimeError
+from ..interfaces import IAdapter, IOutput
 from ..tools.log_helper import ErrorLogger, loggable
 from .input import Input
 from .output import Output